refactor(ntx-builder): extract transaction selection out of the account actor
