+++ bin/ntx-builder/src/candidate.rs
